+++ lib/restful_json.rb
@@ -5,5 +5,4 @@ require 'twinturbo/controller'
 require 'restful_json/base_controller'
 require 'restful_json/controller'
 require 'restful_json/default_controller'
-
-ActiveSupport::Dependencies.autoload_paths << "#{Rails.root}/app/permitters" unless ActiveSupport::Dependencies.autoload_paths.include?("#{Rails.root}/app/permitters")
+require 'restful_json/railtie' if defined? ::Rails::Railtie