+++ spec/dummy/app/controllers/barfoos_controller.rb
@@ -1,5 +1,6 @@
 class BarfoosController < ApplicationController
-  include RestfulJson::DefaultController
+  # test deprecated way to include RestfulJson::DefaultController
+  acts_as_restful_json
   
   query_for :some_action, is: ->(t,q) {q.where(:status => 2)}
   serialize_action :some_action, with: SimpleBarfooSerializer